Success and cancel order pages updated
