+++ app/javascript/packs/components/Orders/newOrder.jsx
@@ -1,21 +1,16 @@
 import React, { useState } from 'react';
 import axios from 'axios';
-import { useHistory } from 'react-router-dom';
 import { Button } from 'react-bootstrap';
-import { toast } from 'react-toastify';
-import 'react-toastify/dist/ReactToastify.css';
 import { useCookies } from 'react-cookie';
 import { loadStripe } from '@stripe/stripe-js';
 const stripePromise = loadStripe(
   'pk_test_51HxwzUFr1fOlo3WtjiUKn9CcA3UkYmgYf4EkQ9GBbb8Qk3NCOkLc6htsEoyLV2IG989T7uCkdGfwMUJQszUcHeLq00ZbXFjQum'
 );
 const NewOrder = ({ name, email, postCode, phone, address }) => {
-  let history = useHistory();
   const [cookies] = useCookies();
   let [basket, setBasket] = useState(
     JSON.parse(sessionStorage.getItem('basket'))
   );
-  toast.configure();
 
   let placeOrder = () => {
     let body = createFormBody();
@@ -59,14 +54,6 @@ const NewOrder = ({ name, email, postCode, phone, address }) => {
       .post('/api/v1/orders', body, headers)
       .then(async (res) => await stripe.redirectToCheckout(res.data))
       .catch((err) => console.error(err.message));
-  };
-
-  let confirmStatus = () => {
-    toast('Your order has been sent');
-    sessionStorage.removeItem('basket');
-    setTimeout(() => {
-      history.push('/');
-    }, 500);
   };
 
   return (
