+++ app/javascript/packs/components/checkout/cansel.jsx
@@ -1,7 +1,9 @@
 import React from 'react';
 
 const CheckoutCansel = () => {
-  return <h1>Something whent wrong, continue shopping</h1>;
+  return (
+    <h1 className="text-center">Something whent wrong, continue shopping</h1>
+  );
 };
 
 export default CheckoutCansel;
