+++ app/javascript/packs/components/checkout/success.jsx
@@ -1,6 +1,23 @@
-import React from 'react';
+import React, { useEffect } from 'react';
+import { useHistory } from 'react-router-dom';
+import { toast } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
+
 const CheckoutSuccess = () => {
-  return <h1>Thank you for your purches</h1>;
+  let history = useHistory();
+  toast.configure();
+
+  const confirmStatus = () => {
+    toast('Your order has been sent');
+    sessionStorage.removeItem('basket');
+    setTimeout(() => {
+      history.push('/');
+    }, 2000);
+  };
+
+  useEffect(() => confirmStatus(), []);
+
+  return <h1 className="text-center">Thank you for your purches</h1>;
 };
 
 export default CheckoutSuccess;
